// admin-dashboard/src/app/lib/types/utils.d.ts
export type PseudoNum = '0' | '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9';